+++ app/javascript/components/game/map/MapHexPatterns.tsx
@@ -50,7 +50,12 @@ export default function MapHexPatterns({ map }: MapHexPatternProps) {
   const nations = [
     "ger", "ita", "jap", "fin", "bul", "hun", "rom", "slv", "cro",
     "uk", "usa", "ussr", "chi", "fra",
-    "frf", "sa", "can", "aus", "nz", "ind", "bra", "pol", "bel", "dut", "nor", "yug", "gre",    
+    "frf", "sa", "can", "aus", "nz", "ind", "bra", "pol", "bel", "dut", "nor", "yug", "gre",
+    "rsp", "eth", "bol", "chb",
+    "nsp", "chc", "chb2", "chg", "par",
+    "sk", "un", "doi", "isr",
+    "nk", "pak", "vie", "syr", "jor", "egy",
+    "cze",
   ]
 
   return (